implemented a version including all improvements that were made to the session wizard
