+++ package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@corbado/webcomponent": "^1.0.34",
+    "@corbado/webcomponent": "^1.0.104",
     "@types/node": "20.2.5",
     "@types/react-dom": "18.2.4",
     "axios": "^1.4.0",
